refactor : Member 검증 도메인 서비스 위치 변경
