+++ src/main/java/makeus/cmc/malmo/application/service/helper/member/MemberQueryHelper.java
@@ -62,4 +62,20 @@ public class MemberQueryHelper {
         }
     }
 
+    public void isMemberCouple(MemberId memberId) {
+        boolean coupleMember = validateMemberPort.isCoupleMember(memberId);
+
+        if (!coupleMember) {
+            throw new NotCoupleMemberException("커플 등록 전인 사용자입니다. 커플 등록 후 이용해주세요.");
+        }
+    }
+
+    public void isMemberValid(MemberId memberId) {
+        boolean validMember = validateMemberPort.isValidMember(memberId);
+
+        if (!validMember) {
+            throw new MemberNotFoundException("존재하지 않는 사용자입니다. 회원가입 후 이용해주세요.");
+        }
+    }
+
 }
